added test coverage report for #17 joined all test into one for coverage to work better added istambul package added istambul command with html and text report
